requirements.txt
# Core dependencies for 2026 Summer Internship pipeline

# Crew.ai and AI/ML
crewai>=0.28.0
langchain>=0.1.0
langchain-openai>=0.1.0
openai>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
requests>=2.31.0

# Database
sqlite3  # Built-in with Python
psycopg2-binary>=2.9.0  # For PostgreSQL support

# Data analysis and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Text processing and NLP
nltk>=3.8.0
spacy>=3.6.0
textblob>=0.17.0

# Web scraping and API
beautifulsoup4>=4.12.0
scrapy>=2.10.0
httpx>=0.24.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.65.0
schedule>=1.2.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Logging and monitoring
structlog>=23.1.0
sentry-sdk>=1.32.0

# Configuration management
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Async support
asyncio
aiohttp>=3.8.0
aiosqlite>=0.19.0

# Data validation
cerberus>=1.3.0
jsonschema>=4.19.0

# Caching
redis>=4.6.0
diskcache>=5.6.0

# File handling
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Date/time handling
python-dateutil>=2.8.0
pytz>=2023.3

# Environment variables
python-decouple>=3.8

# API documentation
fastapi>=0.100.0
uvicorn>=0.23.0

# Background tasks
celery>=5.3.0
redis>=4.6.0

# Email notifications
sendgrid>=6.10.0

# Slack notifications
slack-sdk>=3.21.0

# Data export
tabulate>=0.9.0
rich>=13.5.0
